+++ ConsoleApp1/Program/Movies.cs
@@ -131,41 +131,6 @@ namespace CinemaApplication
                 .Selectable(ConsoleColor.White, ConsoleColor.DarkGray)
                 .LinkWindows(movieWindows.ToArray())
                 .Result();
-            /*
-            if (!root[movieNumber].Equals(null))
-            {
-                Console.ForegroundColor
-                 = ConsoleColor.Green;
-                Console.WriteLine(
-                    $"{root[movieNumber].GetProperty("name")}\n\n" +
-                    $"   Duration: {root[movieNumber].GetProperty("duration")} minutes\n" +
-                    $"   Release Date: {root[movieNumber].GetProperty("releaseDate")}\n" +
-                    $"   Rating: {root[movieNumber].GetProperty("rating")}\n" +
-                    $"   Language: {root[movieNumber].GetProperty("language")}\n" +
-                    $"   Company: {root[movieNumber].GetProperty("company")}"
-                    );
-
-                Console.WriteLine("   Genre(s):");
-                foreach (JsonElement genres in root[movieNumber].GetProperty("genres").EnumerateArray())
-                {
-                    Console.WriteLine($"\t{genres}");
-                }
-
-                Console.WriteLine("   Starring:");
-                foreach (JsonElement starring in root[movieNumber].GetProperty("starring").EnumerateArray())
-                {
-                    Console.WriteLine($"\t{starring}");
-                }
-
-                Console.WriteLine("   Description:");
-                Console.WriteLine($"   {root[movieNumber].GetProperty("description")}");
-            }
-            else
-                Console.WriteLine($"Film {movieNumber} does not exist!");
-           
-            Console.ForegroundColor
-             = ConsoleColor.White;
-            */
         }
     }
